Renames identity to entity
