+++ src/api/entity/requests.rs
@@ -1,16 +1,16 @@
-use super::responses::CreateIdentityResponse;
+use super::responses::CreateEntityResponse;
 use rustify_derive::Endpoint;
 use std::fmt::Debug;
 
 #[derive(Builder, Debug, Default, Endpoint)]
 #[endpoint(
     path = "identity/entity",
-    response = "CreateIdentityResponse",
+    response = "CreateEntityResponse",
     method = "POST",
     builder = "true"
 )]
 #[builder(setter(into), default)]
-pub struct CreateIdentityRequest {
+pub struct CreateEntityRequest {
     pub name: String,
     pub policies: String,
 }
