+++ src/api/entity_alias/requests.rs
@@ -1,4 +1,4 @@
-use super::responses::CreateIdentityAliasResponse;
+use super::responses::CreateEntityAliasResponse;
 use rustify_derive::Endpoint;
 use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
@@ -6,12 +6,12 @@ use std::fmt::Debug;
 #[derive(Builder, Debug, Default, Endpoint, Deserialize, Serialize)]
 #[endpoint(
     path = "identity/entity-alias",
-    response = "CreateIdentityAliasResponse",
+    response = "CreateEntityAliasResponse",
     method = "POST",
     builder = "true"
 )]
 #[builder(setter(into), default)]
-pub struct CreateIdentityAliasRequest {
+pub struct CreateEntityAliasRequest {
     pub name: String,
     pub canonical_id: String,
     pub mount_accessor: String,
