+++ src/entity.rs
@@ -1,5 +1,5 @@
-use crate::api::identity::requests::CreateIdentityRequest;
-use crate::api::identity::responses::CreateIdentityResponse;
+use crate::api::entity::requests::CreateEntityRequest;
+use crate::api::entity::responses::CreateEntityResponse;
 use crate::{
     api::{self},
     client::Client,
@@ -7,12 +7,12 @@ use crate::{
 };
 
 #[instrument(skip(client), err)]
-pub async fn create_identity(
+pub async fn create_entity(
     client: &impl Client,
     name: &str,
     policies: &str,
-) -> Result<CreateIdentityResponse, ClientError> {
-    let endpoint = CreateIdentityRequest::builder()
+) -> Result<CreateEntityResponse, ClientError> {
+    let endpoint = CreateEntityRequest::builder()
         .name(name)
         .policies(policies)
         .build()
