+++ src/entity_alias.rs
@@ -1,16 +1,16 @@
 use crate::{api, client::Client, error::ClientError};
 
-use crate::api::identity_alias::requests::CreateIdentityAliasRequest;
-use crate::api::identity_alias::responses::CreateIdentityAliasResponse;
+use crate::api::entity_alias::requests::CreateEntityAliasRequest;
+use crate::api::entity_alias::responses::CreateEntityAliasResponse;
 
 #[instrument(skip(client), err)]
-pub async fn create_identity_alias(
+pub async fn create_entity_alias(
     client: &impl Client,
     name: &str,
     canonical_id: &str,
     mount_accessor: &str,
-) -> Result<CreateIdentityAliasResponse, ClientError> {
-    let endpoint = CreateIdentityAliasRequest::builder()
+) -> Result<CreateEntityAliasResponse, ClientError> {
+    let endpoint = CreateEntityAliasRequest::builder()
         .name(name)
         .canonical_id(canonical_id)
         .mount_accessor(mount_accessor)
