+++ src/lib.rs
@@ -199,8 +199,8 @@ pub mod auth;
 pub mod client;
 pub mod database;
 pub mod error;
-pub mod identity;
-pub mod identity_alias;
+pub mod entity;
+pub mod entity_alias;
 pub mod kv2;
 pub mod pki;
 pub mod ssh;
